code-guardian-evaluation/requirements.txt
# Code Guardian VS Code Extension Evaluation Requirements
# Python: 3.13.x

# Core data analysis and statistics
numpy>=2.1,<3
pandas>=2.2,<3
scipy>=1.14,<2
scikit-learn>=1.5,<2
matplotlib>=3.9,<4
seaborn>=0.13,<0.14

# Statistical analysis for academic evaluation
statsmodels>=0.14.1,<0.15
pingouin>=0.5,<0.6

# VS Code automation and testing
selenium>=4.18,<5
webdriver-manager>=4,<5

# Code analysis and processing
pygments>=2.17,<3
astunparse>=1.6.3,<2

# Data management for evaluation results
sqlalchemy>=1.4,<2
tinydb>=4.8,<5

# CLI and utilities
click>=8.1,<9
tqdm>=4.66,<5
rich>=13,<14
typer>=0.12,<1

# Configuration management
pydantic>=2.7,<3
python-dotenv>=1.0,<2
pyyaml>=6.0.1,<7

# Testing framework
pytest>=8,<9
pytest-cov>=4,<6

# Visualization for results
plotly>=5,<6

# File I/O for datasets and results
openpyxl>=3.1,<4
jsonlines>=3.1,<4

# Security tools integration (for baseline comparison)
bandit>=1.7,<2
semgrep>=1.64

# Code quality tools (optional, for test code generation)
black>=24,<25
isort>=5.12,<6

# Web utilities for downloading datasets
requests>=2.31,<3

# Time measurement precision
psutil>=5.9,<6

# Network communication for extension testing
websockets>=12,<13
